pyproject.toml
[project]
name = "harry-potter-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix>=13.15.0",
    "ddgs>=9.11.4",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "openai>=2.28.0",
    "openinference-instrumentation-langchain>=0.1.61",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
]
